+++ Sources/KeyboardKit/Previews/KeyboardActionHandler+Preview.swift
@@ -9,7 +9,8 @@
 import CoreGraphics
 
 /**
- This class can be used to preview keyboard views.
+ This class can be used to preview keyboard views. Don't use
+ it in other situations.
  */
 public class PreviewKeyboardActionHandler: KeyboardActionHandler {